scripts/update_anki_template/pyproject.toml
[tool.poetry]
name = "update-anki-template"
version = "0.1.0"
description = ""
authors = ["Félix JULLIEN <jullienfelix@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
anki = "^24.4.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
